--- apps/app/src/app/@common/location/ui/LocationSelect/ListContainer/Data/index.ts
@@ -1,1 +1,0 @@
-export { Data } from "./Data";
