+++ apps/app/src/app/@common/location/ui/LocationSelect/ListContainer/ListContainer.tsx
@@ -1,14 +1,23 @@
 import type { Fulltext } from "@use-pico/client/ui/fulltext";
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
+import { SpinnerContainer } from "@use-pico/client/ui/container";
+import { Button } from "@use-pico/client/ui/button";
+import { withFallback } from "@use-pico/client/utils";
+import { useLocale } from "@use-pico/client/hook";
+import { Container } from "@use-pico/client/ui/container";
+import type { tLocation } from "@zbav-se.me/sdk/api/session";
+import { withLocationAutocompleteQuery } from "@zbav-se.me/sdk/query/session";
 import { Default } from "./Default";
-import { Pending } from "./Pending";
+import { Empty } from "./Data/Empty";
+import { uiSelectButton } from "@zbav-se.me/ui/ui";
 
 export namespace ListContainer {
 	export interface Props
-		extends Omit<Data.Props, "_suspense" | "text">,
+		extends Omit<Container.Props, "onChange">,
 			Pick<Default.Props, "textHint" | "warningStatusProps"> {
 		search: Fulltext.Value;
+		value: string | undefined | null;
+		onChange(value: string): void;
+		onLocation?(value: tLocation): void;
 	}
 }
 
@@ -18,14 +27,16 @@ export namespace ListContainer {
  *
  * @see apps/app/src/app//draft/ui/DraftEditor/DraftEditor.tsx
  */
-export const ListContainer: FC<ListContainer.Props> = ({
+export const ListContainer = withFallback(({
 	textHint,
 	search,
 	value,
+	onChange,
+	onLocation,
 	ui,
 	warningStatusProps,
 	...props
-}) => {
+}: ListContainer.Props) => {
 	const text = search ?? value ?? "";
 
 	if (text.length < 3) {
@@ -38,15 +49,55 @@ export const ListContainer: FC<ListContainer.Props> = ({
 		);
 	}
 
+	const locale = useLocale();
+	const { data } = withLocationAutocompleteQuery.useSuspenseQuery({
+		lang: locale,
+		text,
+	});
+
+	if (data.length === 0) {
+		return <Empty ui={ui} />;
+	}
+
 	return (
-		<Suspense fallback={<Pending />}>
-			<Data
-				_suspense={"I know"}
-				text={text}
-				value={value}
-				ui={ui}
-				{...props}
-			/>
-		</Suspense>
+		<Container
+			data-ui="ListContainer[Container.content]"
+			ui={{
+				scroll: "vertical",
+				height: "full",
+				...ui,
+			}}
+			{...props}
+		>
+			<Container
+				ui={{
+					layout: "vertical-flex",
+					gap: "default",
+				}}
+			>
+				{data.map((item) => {
+					return (
+						<Button
+							key={item.id}
+							onClick={() => {
+								onChange(item.id);
+								onLocation?.(item);
+							}}
+							truncate
+							{...uiSelectButton({
+								isSelected: value === item.id,
+								ui,
+								className: [
+									"text-left",
+								],
+							})}
+							data-ui="ListContainer-[Button]"
+						>
+							{item.address}
+						</Button>
+					);
+				})}
+			</Container>
+		</Container>
 	);
-};
+}, SpinnerContainer);
