+++ apps/app/src/app/@common/location/ui/LocationValue/LocationValue.tsx
@@ -1,10 +1,10 @@
-import { LabelValue } from "@use-pico/client/ui/container";
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { LabelValue, SpinnerContainer } from "@use-pico/client/ui/container";
+import { withFallback } from "@use-pico/client/utils";
+import { withLocationFetchQuery } from "@zbav-se.me/sdk/query/session";
 
 export namespace LocationValue {
-	export interface Props extends LabelValue.PropsEx {
+	export interface Props extends LabelValue.PropsEx, MarkSuspense.Props {
 		locationId: string | undefined | null;
 	}
 }
@@ -15,24 +15,42 @@ export namespace LocationValue {
  *
  * @see apps/app/src/app//draft/ui/DraftEditor/DraftEditor.tsx
  */
-export const LocationValue: FC<LocationValue.Props> = ({ locationId, ...props }) => {
-	if (!locationId) {
+export const LocationValue = withFallback(
+	({ _suspense, locationId, ...props }: LocationValue.Props) => {
+		if (!locationId) {
+			return (
+				<LabelValue
+					data-ui={"LocationValue[LabelValue.empty]"}
+					{...props}
+					textValue={null}
+				/>
+			);
+		}
+
+		const { data } = withLocationFetchQuery.useSuspenseQuery({
+			where: {
+				id: locationId,
+			},
+		});
+
 		return (
 			<LabelValue
-				data-ui={"LocationValue[LabelValue.empty]"}
 				{...props}
-				textValue={null}
+				textValue={data.address}
 			/>
 		);
-	}
-
-	return (
-		<Suspense fallback={<Pending {...props} />}>
-			<Data
-				_suspense={"I know"}
-				locationId={locationId}
+	},
+	({ ...props }: Omit<LocationValue.Props, "_suspense">) => {
+		return (
+			<LabelValue
+				textValue={
+					<SpinnerContainer
+						type="icon"
+						size="md"
+					/>
+				}
 				{...props}
 			/>
-		</Suspense>
-	);
-};
+		);
+	},
+);
